Add CodeMirror editor
