+++ web/codelabs.html
@@ -71,7 +71,7 @@
         </div>
         <div id="right-panel">
             <div id="editor-panel">
-
+                <div id="editor-host"></div>
             </div>
             <div id="output-panel">
